source/lambdas/requirements.txt
boto3
DotMap
Pillow
XlsxWriter
pandas
# amazon-textract-caller
# amazon-textract-response-parser
pyrhubarb